address comment, fix a bug of substring
